COM-50: Handshake Requests need to be decoupled.
First implementation, trying to decouple purely via the COM server. If this
does not work out (e. G. reentrancy problems in Outlook), we need to pass the
callback down to outlook, and open a pseudo-modal dialog there (which blocks
the main window, but does not block the code like ShowDialog()).